Remove duplicate mock and fixture data, re-organise storybook
