+++ packages/storybook/src/ui/organisms/questionnaire-table/questionnaire-table.stories.tsx
@@ -1,0 +1,12 @@
+import { Story } from '@storybook/react'
+import { QuestionnaireTable } from '@ltht-react/table'
+import { summaryDefinition, summaryRecordsList } from './questionnaire-table.fixtures'
+
+export const VerticalTable: Story = () => (
+  <QuestionnaireTable definition={summaryDefinition} records={summaryRecordsList} />
+)
+export const HorizontalTable: Story = () => (
+  <QuestionnaireTable definition={summaryDefinition} records={summaryRecordsList} headerAxis="x" />
+)
+
+export default { title: 'UI/Organisms/QuestionnaireTable' }
